tokit/requirements_extra.txt
# If use SQL
sqlbuilder>=0.7,<0.8

# PostgreSQL
momoko>=2.2.1,<3.0.0

# If use Cassandra
cassandra-driver>=2.7.2,<3.0.0


# If use Coffee and SASS
coffeescript==1.1.2
libsass==0.10.1

oauth2

